Feat/PRSD-1220 restrict local authorities to england only (#496)
